[SeoBundle] Add redirect target path resolution
When reading redirects with php code the target can be either a path or a site id.
The type of the redirect does not reflect the nature of the target link.
With type `path` it can either be an id or a path string.
Therefore an additional method for getting the actual target path is required.